+++ src/index.ts
@@ -30,10 +30,12 @@ import * as path from 'path';
 import psTree from 'ps-tree';
 import { gt } from 'semver';
 import { Mapping, RawSourceMap, SourceMapGenerator } from 'source-map';
+import WebSocket from 'ws';
 import { Logger } from './logger';
 import chalk = require('chalk');
 import rimraf = require('rimraf');
 import acorn = require('acorn');
+import EventEmitter = require('events');
 
 const DEFAULT_TEMP_DIR_NAME = '.dev-server';
 const CORE_MODULE = 'iobroker.js-controller';
@@ -60,6 +62,8 @@ class DevServer {
   private tempDir!: string;
   private profileName!: string;
   private profileDir!: string;
+
+  private readonly socketEvents = new EventEmitter();
 
   constructor() {
     const parser = yargs(process.argv.slice(2));
@@ -512,9 +516,10 @@ class DevServer {
     );
 
     // admin proxy
+    const hiddenAdminPort = this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET);
     app.use(
       createProxyMiddleware([`!${adminPattern}`, '!/browser-sync/**'], {
-        target: `http://localhost:${this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET)}`,
+        target: `http://localhost:${hiddenAdminPort}`,
         ws: true,
       }),
     );
@@ -523,8 +528,10 @@ class DevServer {
     this.log.notice(`Starting web server on port ${config.adminPort}`);
     const server = app.listen(config.adminPort);
 
+    let exiting = false;
     process.on('SIGINT', () => {
       this.log.notice('dev-server is exiting...');
+      exiting = true;
       server.close();
       // do not kill this process when receiving SIGINT, but let all child processes exit first
     });
@@ -534,6 +541,41 @@ class DevServer {
       server.on('error', reject);
       server.on('close', reject);
     });
+
+    const connectWebSocketClient = (): void => {
+      if (exiting) return;
+      // TODO: replace this with @iobroker/socket-client
+      const client = new WebSocket(`ws://localhost:${hiddenAdminPort}/?sid=${Date.now()}&name=admin`);
+      client.on('open', () => this.log.debug('WebSocket open'));
+      client.on('close', () => {
+        this.log.debug('WebSocket closed');
+        setTimeout(connectWebSocketClient, 1000);
+      });
+      client.on('error', (error) => this.log.debug(`WebSocket error: ${error}`));
+      client.on('message', (msg) => {
+        if (typeof msg === 'string') {
+          try {
+            const data = JSON.parse(msg);
+            if (!Array.isArray(data) || data.length === 0) return;
+            switch (data[0]) {
+              case 0:
+                if (data.length > 3) {
+                  this.socketEvents.emit(data[2], data[3]);
+                }
+                break;
+              case 1:
+                // ping received, send pong (keep-alive)
+                client.send('[2]');
+                break;
+            }
+          } catch (error) {
+            this.log.error(`Couldn't handle WebSocket message: ${error}`);
+          }
+        }
+      });
+    };
+
+    connectWebSocketClient();
 
     this.log.box(`Admin is now reachable under http://localhost:${config.adminPort}/`);
   }
@@ -770,7 +812,7 @@ class DevServer {
     const adapterRunDir = path.join(this.profileDir, 'node_modules', `iobroker.${this.adapterName}`);
     await this.startFileSync(adapterRunDir);
 
-    this.startNodemon(adapterRunDir, pkg.main);
+    await this.startNodemon(adapterRunDir, pkg.main);
   }
 
   private async startTscWatch(): Promise<void> {
@@ -838,7 +880,7 @@ class DevServer {
     });
   }
 
-  private startNodemon(baseDir: string, scriptName: string): void {
+  private async startNodemon(baseDir: string, scriptName: string): Promise<void> {
     const script = path.resolve(baseDir, scriptName);
     this.log.notice(`Starting nodemon for ${script}`);
 
@@ -902,6 +944,15 @@ class DevServer {
           process.exit(-1);
         }
       });
+
+    if (!this.isJSController()) {
+      this.socketEvents.on('objectChange', (args: any) => {
+        if (Array.isArray(args) && args.length > 1 && args[0] === `system.adapter.${this.adapterName}.0`) {
+          this.log.notice(`Adapter configuration changed, restarting nodemon...`);
+          nodemon.restart();
+        }
+      });
+    }
   }
 
   async handleNodemonDetailMsg(message: string): Promise<void> {